climbing: refactor getDifficulty* to take FeatureTags (#391)
